backend/requirements.txt
fastapi==0.104.1

uvicorn[standard]==0.24.0

requests==2.31.0

python-dotenv==1.0.0

google-generativeai==0.3.1

pydantic==2.5.0

python-multipart==0.0.6

aiohttp==3.9.1

tenacity==8.2.3

pytest==7.4.3


